\docType{package}
\name{tools4pkgs-package}
\alias{tools4pkgs-package}
\alias{tools4pkgs}
\title{Tools for Package Administration}
\description{
  Helper functions for extracting package dependencies from DESCRIPTION file and for mirroring subset of CRAN to local R repository.
}
\author{
  Jan Gorecki
}
\references{
  \href{https://svn.r-project.org/R/branches/tools4pkgs/}{tools4pkgs branch in base R}
}
\seealso{
  \code{\link{mirror.packages}}, \code{\link{packages.dcf}}, \code{\link{repos.dcf}}
}